requirements.txt
flask
flask_restful
flask_sqlalchemy
flask_migrate
flask_jwt_extended
python-dotenv
redis
celery
werkzeug
flask_bcrypt
flask_cors
redis 
celery 
flask-caching
reportlab
requests
boto3
croniter
requests
sqlalchemy-utils